add docstring and split into more files
